app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.40
langchain-ollama==0.2.3
langsmith==0.3.11
numpy==2.2.3
ollama==0.4.7
ollama-ocr==0.1.3
openai==1.68.2
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pytesseract==0.3.13
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
sniffio==1.3.1
starlette==0.45.3
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
